archive/pyproject.toml
[project]
name = "asktheworld"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32",
    "openai>=1.75",
    "python-dotenv>=1.0",
    "ollama>=0.4",
]
